+++ src/Dtl3.cc
@@ -35,13 +35,13 @@ Handle<Value> Dtl3::New(const Arguments& args)
 
         if (val->IsInt32()) {
             vector<int> seq0, seq1, seq2;
-            for (int i=0;i<a0->Length();++i) {
+            for (size_t i=0;i<a0->Length();++i) {
                 seq0.push_back(a0->Get(Integer::New(i))->Int32Value());
             }
-            for (int i=0;i<a1->Length();++i) {
+            for (size_t i=0;i<a1->Length();++i) {
                 seq1.push_back(a1->Get(Integer::New(i))->Int32Value());
             }
-            for (int i=0;i<a2->Length();++i) {
+            for (size_t i=0;i<a2->Length();++i) {
                 seq2.push_back(a2->Get(Integer::New(i))->Int32Value());
             }
             Dtl3 *dtl3 = new Dtl3(seq0, seq1, seq2);
@@ -50,13 +50,13 @@ Handle<Value> Dtl3::New(const Arguments& args)
             return args.This();
         } else {
             vector<string> seq0, seq1, seq2;
-            for (int i=0;i<a0->Length();++i) {
+            for (size_t i=0;i<a0->Length();++i) {
                 seq0.push_back(string(*String::Utf8Value(a0->Get(Integer::New(i))->ToString())));
             }
-            for (int i=0;i<a1->Length();++i) {
+            for (size_t i=0;i<a1->Length();++i) {
                 seq1.push_back(string(*String::Utf8Value(a1->Get(Integer::New(i))->ToString())));
             }
-            for (int i=0;i<a2->Length();++i) {
+            for (size_t i=0;i<a2->Length();++i) {
                 seq2.push_back(string(*String::Utf8Value(a2->Get(Integer::New(i))->ToString())));
             }
             Dtl3 *dtl3 = new Dtl3(seq0, seq1, seq2);
@@ -98,14 +98,14 @@ Handle<Value> Dtl3::MergedSequence(const Arguments& args)
     if (dtl3->getType() == DtlTypeStringArray) {
         vector<string> seq = dtl3->vsdiff3->getMergedSequence();
         Local<Array> ret = Array::New(seq.size());
-        for (int i=0;i<seq.size();++i) {
+        for (size_t i=0;i<seq.size();++i) {
             ret->Set(Integer::New(i), String::New(seq[i].c_str()));
         }
         return scope.Close(ret);
     } else if (dtl3->getType() == DtlTypeIntArray) {
         vector<int> seq = dtl3->vidiff3->getMergedSequence();
         Local<Array> ret = Array::New(seq.size());
-        for (int i=0;i<seq.size();++i) {
+        for (size_t i=0;i<seq.size();++i) {
             ret->Set(Integer::New(i), Integer::New(seq[i]));
         }
         return scope.Close(ret);
